requirements.txt
numpy==1.22.3
pandas==1.4.2
pyodbc==4.0.32
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
